+++ Source/ShapeMenu/Views/ColorPickerView/ColorPickerView.xib
